+++ packages/pluggableWidgets/color-picker-web/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Fixed
+
+-   We fixed an issue with Color Picker not working in React client.
+
 ## [2.1.2] - 2024-08-28
 
 ### Changed